Pass tgt-client to sign_authdata if for_user_princ is local
In protocol transition, when processing the first TGS-REQ (client realm),
the PAC in the TGT belongs to tgt-client, that is the service and not to
for_user_princ. We need to pass it to sign_authdata so it can verity the
PAC, while for signing it can use client->princ (which matches for_user).

XXX: Update doc.

Signed-off-by: Isaac Boukris <iboukris@gmail.com>