Use file2 replay cache by default
Remove the existing default replay cache implementation and replace it
with a wrapper around the file2 replay cache code.  Change the
filename to krb5_EUID.rcache2, ignoring the residual (and therefore
the server principal name).  On Windows, use the local appdata
directory if KRB5RCACHEDIR is not set in the environment.

ticket: 8786