Make rcache file2 code slightly more robust
In rc_file2.c:store(), when making note of available records,
explicitly check for an empty record (r1stamp or t2stamp is 0), to
more closely match the check for terminating the search.  This
silences a Coverity false positive, as Coverity does not assume that
now > skew as it would be in practice.

To preserve code readability, shorten some variable names, add an
expired() inline helper function, and add comments.