Clean up default ccache in dejagnu test suite
The dejagnu test suite could fail due to a leftover tmpdir/tkt ccache
containing expired tickets.  Remove the ccache file along with the
keytab in delete_db.  (The file cpw_keytab, formerly written as
cpw_srvtab, is not used in the dejagnu tests and has not been used
going back to 1.0, so replace that removal with the removal of
tmpdir/tkt.)