Improve daemon checking in Python test scripts
If a daemon exits early and we detect it with check_daemon(), avoid
trying to terminate it again as the process entry will have been
reaped.  Check all daemons on successful exit and exit with an error
if any daemons exited early.

Also remove a piece of Python 2.5 compatibility code which is no
longer relevant with Python 3.