Add minimal cross-realm support to test KDB module
Cross-realm is already possible with the test KDB module, but because
the lookup realm is not included in the profile query or key
derivation, inbound cross-TGTs are implicit and use the same keys as
the local TGT, potentially obscuring bugs.  Add the lookup realm to
key derivation so that they use different keys.