Simplify AS request time handling in KDC
The kdc_time and authtime fields of struct as_req_state are redundant
and can be condensed to just kdc_time.  Copying the times structure
from enc_tkt_reply to reply_encpart already sets the authtime field to
kdc_time, so there is no need to repeat that assignment.

Also remove two prototypes for functions which never existed in the
mainline KDC code.