Improve type safety of serialization code
Remove the serialization table from krb5_context, the functions to
find and register serializers, and the polymorphic serialization
functions.  Instead, directly call per-type serialization functions
for the type of object we need to serialize.

Remove the krb5_context parameter from most serialization functions as
the interfaces are not public and the context is not needed.

Remove the ccache, keytab, and rcache serialization handlers as they
were not used.

In t_ser.c, repeat the externalize-internalize-reexternalize logic for
each type, but simplify the code by aborting on failure and removing
the verbose flag.

The krb5_context and krb5_keyblock sizing functions both reported an
extra four bytes.  The new test program does not tolerate that
inconsistency, so fix them.