Fix KDC crash in handle_signticket
Regression since RBCD changes, commit d47f7dba3779 changed the
parameters passed to sign_authdata() in S4U2Proxy so that client
is the impersonted client (used to be impersonator), and added
header_server for the impersonator.  But it failed to fix the call
to handle_signticket() to use header_server instead of client which
now can be NULL if the client is not local.