Document enctype migration
In enctypes.rst, note that des3-cbc-sha1 and arcfour-hmac are
deprecated, and add a paragraph about des3-cbc-sha1 support being
removed in the future.  Add a section on migrating away from legacy
enctypes (briefer than the existing advanced/retiring-des.rst and not
specific to single-DES).

In kdc_conf.rst, note that des3-cbc-sha1 and arcfour-hmac are
deprecated.

ticket: 8950 (new)