cc: add krb5_cc_notification_path
Adds a new API call krb5_cc_notification_path that implements
generic notification interface. The function returns a path
to the file that is touched everytime the ccache or collection
is modified.

The caller can setup a filesystem watch to watch for the events.
For example inotify's IN_ATTRIB event.

This also adds a new configure flags --with-notification-path. This
points to the path that will hold the notification file, the path
must be readable only by the owner of the ccache. If not set, the
functionality will be disabled.

It is recommended to set it on the user runtime directory when
available, i.e. --with-notification-path=/run/user/%{euid}.