otp: add support for otp-challenge mechanism
otp-challenge flag requires edata to obtain a challenge message from
the server that must be presented to the user in order to authenticate
correctly.

In order to do so, otp_edata sends a password-less Access-Request and
expects Access-Challenge in reply. Then reads the challenge from
Reply-Message attribute and remembers the State attribute inside a
cookie.

Then it reads the cookie in otp_verify flow and appends it to the next
Access-Request together with the authentication token so the RADIUS
server can match the second message with the initial one.