Run Github Actions on ubuntu-latest (not 18.04)
The ubuntu-18.04 image no longer works on Github Actions.  Run most
jobs on ubuntu-latest instead.

libresolv-wrapper does not work on ubuntu-22.04 at this time
(Launchpad bug #2015570), so don't install it for now.  The URI
discovery tests will be skipped until this is rectified.

For the doc build, move doc-newest-sphix to ubuntu-latest; the Ubuntu
version doesn't matter much for this build as it downloads sphinx from
pypi.org.  Move the doc-older-sphinx to ubuntu-22.04 (currently the
same as ubuntu-latest), skipping ubuntu-20.04 as its version of
doxygen (1.18.17) doesn't correctly the type declaration for
krb5_const_principal.