Fix OpenSSL 1.0 and 1.1 support in PKINIT
Commit f745c9a9bd6c0c73b944182173f1ac305d03dc3a uses ECDSA_SIG_set0(),
which was added in OpenSSL 1.1.  Add a compatibility version for
OpenSSL 1.0.

Commit bdcd6075bd4593c8f67722ce075c9519faec58b7 uses
EVP_PKEY_get_base_id(), which was added in OpenSSL 3.0.  Add a
compatibility macro to use the old name for OpenSSL 1.0 and 1.1.

Commit 0f870b1bcad960fd5319a3f97aafd7f4a289e2fb added ECDH support,
but did not change the OpenSSL 1.0 versions of encode_spki(),
decode_spki(), or generate_dh_pkey() to work with elliptic curve
public keys.  In each function, check the key type and skip the
DH-specific handling for key types other than DH.