Improve profile final flag support
When parsing a file, ignore sections appearing after a final-flagged
section of the same name.  Adjust the meaning of group_level in the
parser state so that it is 1 inside of top-level sections instead of
0, and simplify the addition of top-level sections to the tree by
relying on profile_add_node()'s section merging.

Make the final flag work for relations as well as sections.  Check it
while parsing via a new check_final parameter in profile_add_node(),
and during iteration.

Output final flags for relations in dump_profile().  Make the final
flag available to it via a new output parameter in
profile_find_node_relation().

ticket: 9120