Do not reload a modified profile data object
The profile library normally attempts to reload a profile data tree if
the backing file has changed.  Reloading a dirty profile object
discards any modifications made by the caller.  If we assume that the
modifications are destined to be flushed back out to the backing file,
then there is no good answer--one or the other set of changes will be
lost.  But the caller may have a different intended use for the
modified tree (profile_flush_to_file(), profile_flush_to_buffer(),
krb5_init_context_profile()), for which the caller's modifications may
be critical.  Avoid discarding in-memory edits to ensure the
correctness of these use cases.

ticket: 9118